introduce node_utils_k8s and move k8s utils there
