+++ simplyblock_web/blueprints/snode_ops_k8s.py
@@ -16,8 +16,9 @@ from kubernetes.client import ApiException
 from jinja2 import Environment, FileSystemLoader
 import yaml
 
-from simplyblock_web import utils, node_utils
+from simplyblock_web import utils, node_utils, node_utils_k8s
 from simplyblock_core import scripts, constants, shell_utils
+from simplyblock_web.node_utils_k8s import deployment_name, namespace_id_file, pod_name
 
 logger = logging.getLogger(__name__)
 logger.setLevel(constants.LOG_LEVEL)
@@ -25,19 +26,11 @@ bp = Blueprint("snode", __name__, url_prefix="/snode")
 
 cluster_id_file = "/etc/foundationdb/sbcli_cluster_id"
 
-node_name = os.environ.get("HOSTNAME")
-deployment_name = f"snode-spdk-deployment-{node_name}"
-default_namespace = 'default'
-namespace_id_file = '/etc/simplyblock/namespace'
-pod_name = deployment_name[:50]
-
-
 config.load_incluster_config()
 k8s_apps_v1 = client.AppsV1Api()
 k8s_core_v1 = client.CoreV1Api()
 
 TOP_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-# spdk_deploy_yaml = os.path.join(TOP_DIR, 'static/deploy_spdk.yaml')
 
 
 def set_namespace(namespace):
@@ -175,52 +168,11 @@ def get_info():
 
 @bp.route('/join_swarm', methods=['POST'])
 def join_swarm():
-    # data = request.get_json()
-    # cluster_ip = data['cluster_ip']
-    # cluster_id = data['cluster_id']
-    # join_token = data['join_token']
-    # db_connection = data['db_connection']
-    #
-    # logger.info("Setting DB connection")
-    # scripts.set_db_config(db_connection)
-    # set_cluster_id(cluster_id)
-    #
-    # logger.info("Joining Swarm")
-    # node_docker = get_docker_client()
-    # if node_docker.info()["Swarm"]["LocalNodeState"] == "active":
-    #     logger.info("Node is part of another swarm, leaving swarm")
-    #     node_docker.swarm.leave(force=True)
-    #     time.sleep(2)
-    # node_docker.swarm.join([f"{cluster_ip}:2377"], join_token)
-    # retries = 10
-    # while retries > 0:
-    #     if node_docker.info()["Swarm"]["LocalNodeState"] == "active":
-    #         break
-    #     logger.info("Waiting for node to be active...")
-    #     retries -= 1
-    #     time.sleep(1)
-    # logger.info("Joining docker swarm > Done")
-    #
-    # try:
-    #     nodes = node_docker.containers.list(all=True)
-    #     for node in nodes:
-    #         if node.attrs["Name"] == "/spdk_proxy":
-    #             node_docker.containers.get(node.attrs["Id"]).restart()
-    #             break
-    # except:
-    #     pass
-
     return utils.get_response(True)
 
 
 @bp.route('/leave_swarm', methods=['GET'])
 def leave_swarm():
-    # delete_cluster_id()
-    # try:
-    #     node_docker = get_docker_client()
-    #     node_docker.swarm.leave(force=True)
-    # except:
-    #     pass
     return utils.get_response(True)
 
 
@@ -315,12 +267,6 @@ if CLOUD_INFO:
     SYSTEM_ID = CLOUD_INFO["id"]
 else:
     SYSTEM_ID, _, _ = node_utils.run_command("dmidecode -s system-uuid")
-
-
-
-
-
-
 
 
 @bp.route('/spdk_process_start', methods=['POST'])
@@ -481,15 +427,15 @@ def firewall_set_port():
     action = data['action']
     container = "spdk-container"
 
-    resp = k8s_core_v1.list_namespaced_pod(node_utils.get_namespace())
+    resp = k8s_core_v1.list_namespaced_pod(node_utils_k8s.get_namespace())
     for pod in resp.items:
         if pod.metadata.name.startswith(pod_name):
-            ret = node_utils.firewall_port_k8s(port_id, port_type, action=="block", k8s_core_v1, node_utils.get_namespace(), pod.metadata.name, container)
+            ret = node_utils_k8s.firewall_port_k8s(port_id, port_type, action=="block", k8s_core_v1, node_utils_k8s.get_namespace(), pod.metadata.name, container)
             return utils.get_response(ret)
     return utils.get_response(False)
 
 
 @bp.route('/get_firewall', methods=['GET'])
 def get_firewall():
-    ret = node_utils.firewall_get_k8s()
+    ret = node_utils_k8s.firewall_get_k8s()
     return utils.get_response(ret)
